backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
langchain-text-splitters
faiss-cpu
sentence-transformers
docx2txt
sqlalchemy
slowapi
arize-phoenix
